Skip junctions when pruning out recursive vendor dirs
